Move data files to more reasonable place
