+++ website/src/data/members.js
@@ -1,26 +1,26 @@
 // --- Board ---
-import michellelin from '../../../images/ourTeam/memberImages/michellelin.jpg';
-import noah from '../../../images/ourTeam/memberImages/noah.png';
-import justin from '../../../images/ourTeam/memberImages/justin.jpg';
-import carly from '../../../images/ourTeam/memberImages/carly.jpg';
-import gayathri from '../../../images/ourTeam/memberImages/gayathri.jpeg';
-import kaila from '../../../images/ourTeam/memberImages/kailalong.jpeg';
-import andrewlee from '../../../images/ourTeam/memberImages/andrewlee.jpg';
-import arjunsivakumar from '../../../images/ourTeam/memberImages/arjunsivakumar.jpg';
-import maithyle from '../../../images/ourTeam/memberImages/maithyle.jpg';
-import kainoa from '../../../images/ourTeam/memberImages/kainoa.jpeg';
-import emmyc from '../../../images/ourTeam/memberImages/emmyc.jpg';
+import michellelin from '../images/ourTeam/memberImages/michellelin.jpg';
+import noah from '../images/ourTeam/memberImages/noah.png';
+import justin from '../images/ourTeam/memberImages/justin.jpg';
+import carly from '../images/ourTeam/memberImages/carly.jpg';
+import gayathri from '../images/ourTeam/memberImages/gayathri.jpeg';
+import kaila from '../images/ourTeam/memberImages/kailalong.jpeg';
+import andrewlee from '../images/ourTeam/memberImages/andrewlee.jpg';
+import arjunsivakumar from '../images/ourTeam/memberImages/arjunsivakumar.jpg';
+import maithyle from '../images/ourTeam/memberImages/maithyle.jpg';
+import kainoa from '../images/ourTeam/memberImages/kainoa.jpeg';
+import emmyc from '../images/ourTeam/memberImages/emmyc.jpg';
 
 // --- Leads ---
-import madhu from '../../../images/ourTeam/memberImages/madhu.jpg';
-import minh from '../../../images/ourTeam/memberImages/minhnguyen.jpg';
-import michaelpien from '../../../images/ourTeam/memberImages/michaelpien.png';
-import dang from '../../../images/ourTeam/memberImages/dang.jpg';
-import joshlipton from '../../../images/ourTeam/memberImages/joshlipton.jpg';
-import michelleblin from '../../../images/ourTeam/memberImages/michelleblin.jpg';
-import ostend from '../../../images/ourTeam/memberImages/ostend.png';
-import selina from '../../../images/ourTeam/memberImages/selina.jpg';
-import rafael from '../../../images/ourTeam/memberImages/rafael.jpg';
+import madhu from '../images/ourTeam/memberImages/madhu.jpg';
+import minh from '../images/ourTeam/memberImages/minhnguyen.jpg';
+import michaelpien from '../images/ourTeam/memberImages/michaelpien.png';
+import dang from '../images/ourTeam/memberImages/dang.jpg';
+import joshlipton from '../images/ourTeam/memberImages/joshlipton.jpg';
+import michelleblin from '../images/ourTeam/memberImages/michelleblin.jpg';
+import ostend from '../images/ourTeam/memberImages/ostend.png';
+import selina from '../images/ourTeam/memberImages/selina.jpg';
+import rafael from '../images/ourTeam/memberImages/rafael.jpg';
 
 // --- Members ---
 // --- ------- ---
